Rename test modules for CI integrity scan
