+++ crates/temper-spec/src/automaton/parser_core_test.rs
