+++ crates/temper-spec/src/automaton/parser_test.rs
@@ -1,10 +1,10 @@
 pub(super) const ORDER_IOA: &str = include_str!("../../../../test-fixtures/specs/order.ioa.toml");
 
-#[path = "parser_test_core.rs"]
+#[path = "parser_core_test.rs"]
 mod core;
-#[path = "parser_test_features.rs"]
+#[path = "parser_features_test.rs"]
 mod features;
-#[path = "parser_test_integrations.rs"]
+#[path = "parser_integrations_test.rs"]
 mod integrations;
-#[path = "parser_test_triggers.rs"]
+#[path = "parser_triggers_test.rs"]
 mod triggers;
