+++ crates/temper-spec/src/automaton/parser_triggers_test.rs
